remove default dtype, device

+++ test/sparsity/test_fast_sparse_training.py
@@ -28,7 +28,9 @@ class TestRuntimeSemiStructuredSparsity(TestCase):
 
         input = torch.rand((128, 128)).half().cuda()
         grad = torch.rand((128, 128)).half().cuda()
-        model = ToyTwoLinearModel(128, 256, 128, device="cuda").half()
+        model = ToyTwoLinearModel(
+            128, 256, 128, device="cuda", dtype=torch.bfloat16
+        ).half()
         model_c = copy.deepcopy(model)
 
         for name, mod in model.named_modules():
@@ -75,7 +77,9 @@ class TestRuntimeSemiStructuredSparsity(TestCase):
 
         input = torch.rand((128, 128)).half().cuda()
         grad = torch.rand((128, 128)).half().cuda()
-        model = ToyTwoLinearModel(128, 256, 128, device="cuda").half()
+        model = ToyTwoLinearModel(
+            128, 256, 128, device="cuda", dtype=torch.bfloat16
+        ).half()
         model_c = copy.deepcopy(model)
 
         for name, mod in model.named_modules():
